backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
stripe==10.12.0
python-dotenv==1.0.1
